requirements.txt
Django==1.9.4
django-appconf==1.0.1
django-bootstrap-form==3.2
django-searchbar==0.1.7
django-user-accounts==1.3.1
django-widget-tweaks==1.4.1
pinax-theme-bootstrap==7.4.0
psycopg2==2.6.1
pytz==2016.1
six==1.10.0
